ui major update
